% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_rchunks.R
\name{comment_rchunks}
\alias{comment_rchunks}
\title{Comment all R chunks.}
\usage{
comment_rchunks()
}
\description{
Comment all R chunks present in the current document.
}
\details{
Does not comment already commented lines.
Does not comment anything if some chunks are opened but not closed (and reciprocally).
}
