% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chunks_lines.R
\name{get_chunks_lines}
\alias{get_chunks_lines}
\title{Exctract R chunks lines numbers.}
\usage{
get_chunks_lines(script)
}
\arguments{
\item{script}{The contents of the document.}
}
\value{
A list of lines numbers. Each element represent the line numbers of a chunk.
Returns \code{NULL} if some chunks are opened but not closed (and reciprocally).
}
\description{
Exctract the R chunks lines numbers as a list.
}
