% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_rchunks.R
\name{remove_rchunks}
\alias{remove_rchunks}
\title{Remove all R chunks.}
\usage{
remove_rchunks()
}
\description{
Remove all R chunks present in the current document.
}
\details{
Does not remove anything if nested chunks are detected or if some chunks are opened but not closed (and reciprocally).
}
