% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncomment_rchunks.R
\name{uncomment_rchunks}
\alias{uncomment_rchunks}
\title{Uncomment all R chunks.}
\usage{
uncomment_rchunks()
}
\description{
Uncomment all commented R chunks present in the current document.
}
\details{
Does not uncomment anything if some chunks are opened but not closed (and reciprocally).
}
