% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chunks_lines.R
\name{get_chunks_lines}
\alias{get_chunks_lines}
\title{Exctract R chunks lines numbers.}
\usage{
get_chunks_lines(script)
}
\value{
The list of lines numbers. Each element represent a chunk.
}
\description{
Exctract R chunks lines numbers.
}
